try lab2 grade

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,8 +3,8 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
@@ -16,12 +16,12 @@ public class DebugExercise1 {
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
         int t2 = 9;
-        int b2 = 4;
+        float b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
-        int t3 = 3;
-        int b3 = 4;
+        float t3 = 3;
+        float b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
     }
